Scale down keycloak and delete the ingress object ahead the upgrade
During the real world upgrade, we have seen that ArgoCD might not prune
the objects ahead, which can cause that the Wildfily keycloak and the
Quarkus keycloak instances run in parallel. This shound't be an issue,
since we migrate to the exact same version, but to be on the safe side
we better turn off the Wilfly instance ahead.

We saw also a conflict generated by the OpenShift router, having two
different named ingress objects with the same hostname. Deleting the
ingress object ahead prevents it.